Revamping the index - Keeping Admin while having to navigate manually to page ATM
Working on the index which will house the main/featured project and working on whether or not to have user log in or just have them subscribed to projects via email.  Keeping admin the same 90 looking page for right now.  It's for me and not for public
