+++ Extra/tutorials I used.txt
@@ -17,3 +17,8 @@ This is from school, not sure if I am going to use this
 function make_links_clickable($text){
     return preg_replace('!(((f|ht)tp(s)?://)[-a-zA-Zа-яА-Я()0-9@:%_+.~#?&;//=]+)!i', '<a href="$1">$1</a>', $text);
 }
+
+side bar navigation
+https://www.youtube.com/watch?v=V0O4pY2xX10
+pie charts
+https://www.youtube.com/watch?v=XEiZNT9bRQo
