+++ helper/footer.php
@@ -3,3 +3,6 @@
 </footer>
 </body>
 </html>
+<?php
+    mysqli_close($conn);
+?>
