+++ style/style.css
@@ -1,120 +1,121 @@
-/* Genral layout and color scheme of the webpage */
-body {
-    background-color: rgb(46, 46, 46);
-    margin: 0px;
-    font-family: Arial;
-
+/* Basic body tags */
+*{
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    text-decoration: none;
+    font-family: 'Roboto', sans-serif;
 }
 
-header
-{
+body{
+    background-color: grey;
+}
+
+section{
+    height: 100%;
+    background-color: honeydew;
+    transition: all 1s;
+}
+
+nav{
+    position: fixed;
+    left: -255px;
+    top: 0;
+    width: 250px;
+    height:100%;
+    background: teal;
+    transition: all 1s;
+}
+
+nav header{
+    font-size: 22px;
+    color: white;
     text-align: center;
-    height: 20%;
+    line-height: 70px;
+    background: rgb(6, 202, 202);
+    user-select: none;
 }
 
-main
-{
-    /* In-case of no content on page */
-    min-height: 20%;
-}
-
-.basic {
-    width: 60%;
-    background-color: rgb(167, 167, 167);
-    margin: auto;
-}
-
-.currentProjectOutline{
-    min-height: 20px;
-    /* border: solid; */
-    border-bottom: solid;
-    width: 65%;
-    margin: auto;
-    float:right;
-}
-
-.currentProjectOutline img{
-    height: 32px;
-    border-right: solid;
-    border-bottom: solid;
-    width: 32px;
-    margin: auto;
-    float: left;
-    padding: 10px;
-    background-color: black;
-}
-
-.currentProjectOutline h3{
-    height: 32px;
-    border-bottom: solid;
-    border-color: black;
-    margin: auto;
-    padding: 10px;
-    text-align: center;
-    background-color: black;
-    color: whitesmoke;
-    align-items: center;
-}
-
-.currentProjectOutline .myProgress{
-    height: 20px;
+nav ul a {
+    display: block;
     width: 100%;
-    background-color: #ddd;
-    margin: auto;
-    float: right;
-}
-
-.currentProjectOutline table{
-    margin: auto;
-}
-
-.currentProjectOutline label{
-    padding-left: 5px;
-    padding-bottom: 5px;
-}
-
-.currentProjectOutline .myBar{
-    height: 20px;
-    background-color: #039127;
-    margin: auto;
-    float: left;
-}
-
-.currentProjectBlog{
-    width: 85%;
-    margin: auto;
-}
-
-#blogTitle{
-    height: 40px;
-    background-color: black;
+    line-height: 65px;
+    font-size: 20px;
     color: whitesmoke;
-    text-align: center;
-    justify-items: center;
-    padding-top: 10px;
+    padding-left: 40px;
+    box-sizing: border-box;
+    border-bottom: 1px solid black;
+    border-top: 1px solid rgba(255, 255, 255 ,.1);
+    transition: .4s ease;
 }
 
-.projectList{
+ul li:hover a {
+    padding-left: 50px;
+}
+
+nav ul a i {
+    margin-right: 16px;
+}
+
+#check{
+    display: none;
+}
+
+label #menuBtn, label #menuCancel{
+    position: absolute;
+    cursor: pointer;
+    background: #042331;
+    border-radius: 3px;    
+}
+
+label #menuBtn{
+    left: 40px;
+    top: 16px;
+    font-size: 35px;
+    color: white;
+    padding: 6px 12px;
+    transition: all .5s;
+}
+
+label #menuCancel{
+    z-index: 1111;
+    left: -215px;
+    top: 7px;
+    font-size: 15px;
+    color: #0a5275;
+    padding: 4px 9px;
+    opacity: 0;
+    transition: all 1s;
+}
+
+#check:checked ~ nav{
+    left: 0;
+}
+
+#check:checked ~ label #menuBtn{
+    left: 300px;
+    opacity: 0;
+    pointer-events: none;
+}
+
+#check:checked ~ label #menuCancel{
+    left: 215px;
+    opacity: 1;
+}
+
+#check:checked ~ section{
+    margin-left: 250px;
+}
+
+.timeTracker{
+    width: 70%;
     margin: auto;
+    height: 30%;
+    background: cornsilk;
+}
+
+#companyLogo {
     text-align: center;
-    /* background-color: blue; */
-    padding-top: 25px;
-}
-
-.projectList thead{
-    background-color: black;
-    color: whitesmoke;
-}
-
-.projectList th{
-    padding-top: 25px;
-    padding-bottom: 25px;
-}
-
-.quickAboutMe{
-    min-height: 200px;
-    width: 20%;
-    float: left;
-    background-color: blue;
-    /* margin: auto; */
+    font-weight: 700;
+    font-size: 22px;
 }
